+++ packages/lint/src/index.ts
@@ -2,6 +2,7 @@ import buttonRules from './rules/button-rules.js';
 import captionRules from './rules/caption-rules.js';
 import containerRules from './rules/container-rules.js';
 import imageRules from './rules/image-rules.js';
+import kickerRules from './rules/kicker-rules.js';
 import listRules from './rules/list-rules.js';
 import linkRules from './rules/link-rules.js';
 import proseRules from './rules/prose-rules.js';
@@ -12,6 +13,7 @@ const rules = {
   ...captionRules.rules,
   ...containerRules.rules,
   ...imageRules.rules,
+  ...kickerRules.rules,
   ...listRules.rules,
   ...linkRules.rules,
   ...proseRules.rules,
@@ -25,6 +27,7 @@ const configs = {
       ...captionRules.configs.recommended.rules,
       ...containerRules.configs.recommended.rules,
       ...imageRules.configs.recommended.rules,
+      ...kickerRules.configs.recommended.rules,
       ...listRules.configs.recommended.rules,
       ...linkRules.configs.recommended.rules,
       ...proseRules.configs.recommended.rules,
@@ -46,6 +49,8 @@ export type {
   CaptionTagAnalysis,
   ContainerClassInfo,
   ContainerTagAnalysis,
+  KickerClassInfo,
+  KickerTagAnalysis,
   ImageClassInfo,
   ImageTagAnalysis,
   ListClassInfo,